+++ assignment1/src/main.ts
@@ -1,5 +1,5 @@
 import { createApp } from 'vue'
-import App from './game.vue'
+import App from './App.vue'
 import router from './router'
 
 import './assets/main.css'
